Added adaptor and service classes

+++ include/sm_queue.h
@@ -22,7 +22,7 @@ typedef struct __attribute__((aligned(SM_WORD))) sm_queue
 
 sm_queue *sm_queue_create(uint32_t event_size,
                           bool K, bool P, bool H,
-                          unsigned num_of_events,
+                          size_t num_of_events,
                           bool synchronized);
 #define SM_QUEUE_CREATE_EMPTY(S) \
     sm_queue_create(0, false, false, false, 0, (S))
